@@ -11,11 +11,11 @@
   "homepage": "http://developer.constantcontact.com",
   "license": "MIT",
   "require": {
-    "php": ">=5.5.0",
-    "guzzlehttp/guzzle": "^6.1.1"
+    "php": ">=7.3",
+    "guzzlehttp/guzzle": "^7.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "4.8.21"
+    "phpunit/phpunit": "^9.0"
   },
   "authors": [
     {
