@@ -5,13 +5,13 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 
-class ActivityServiceUnitTest extends PHPUnit_Framework_TestCase {
+class ActivityServiceUnitTest extends \PHPUnit\Framework\TestCase {
     /**
      * @var Client
      */
     private static $client;
 
-    public static function setUpBeforeClass() {
+    public static function setUpBeforeClass() : void {
         $mock = new MockHandler([
             new Response(200, array(), JsonLoader::getActivity()),
             new Response(200, array(), JsonLoader::getActivities()),
