@@ -13,13 +13,13 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 
-class CampaignTrackingServiceUnitTest extends PHPUnit_Framework_TestCase {
+class CampaignTrackingServiceUnitTest extends \PHPUnit\Framework\TestCase {
     /**
      * @var Client
      */
     private static $client;
 
-    public static function setUpBeforeClass() {
+    public static function setUpBeforeClass() : void {
         $mock = new MockHandler([
             new Response(200, array(), JsonLoader::getBounces()),
             new Response(200, array(), JsonLoader::getClicks()),
