@@ -9,13 +9,13 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 
-class EmailMarketingServiceUnitTest extends PHPUnit_Framework_TestCase {
+class EmailMarketingServiceUnitTest extends \PHPUnit\Framework\TestCase {
     /**
      * @var Client
      */
     private static $client;
 
-    public static function setUpBeforeClass() {
+    public static function setUpBeforeClass() : void {
         $getCampaignJson = JsonLoader::getCampaignJson();
         $mock = new MockHandler([
             new Response(200, array(), JsonLoader::getCampaignsJson()),
