@@ -9,13 +9,13 @@ use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Psr7\Response;
 
-class LibraryServiceUnitTest extends PHPUnit_Framework_TestCase {
+class LibraryServiceUnitTest extends \PHPUnit\Framework\TestCase {
     /**
      * @var Client
      */
     private static $client;
 
-    public static function setUpBeforeClass() {
+    public static function setUpBeforeClass() : void {
         $mock = new MockHandler([
             new Response(200, array(), JsonLoader::getLibraryFileJson()),
             new Response(200, array(), JsonLoader::getLibraryFilesJson()),
